requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
ansicon==1.89.0
anyio==4.9.0
art==6.4
async-tls-client==1.0.9
attrs==25.3.0
backoff==2.2.1
base58==2.1.1
beautifulsoup4==4.13.3
bitarray==3.3.1
blessed==1.20.0
bs4==0.0.2
canoser==0.8.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==2.1.1
colorama==0.4.6
contourpy==1.3.1
curl_cffi==0.10.0
cycler==0.12.1
cytoolz==1.0.1
dataclasses-json==0.6.7
Deprecated==1.2.18
editor==1.6.6
eth-account==0.13.6
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
fake-useragent==2.1.0
fonttools==4.55.3
frozenlist==1.5.0
gql==3.5.2
graphql-core==3.2.4
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hexbytes==1.3.0
hpack==4.1.0
httpcore==1.0.8
httpx==0.27.2
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
inquirer==3.4.0
jinxed==1.3.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
loguru==0.7.3
markdown-it-py==3.0.0
marshmallow==3.26.1
matplotlib==3.10.0
mdurl==0.1.2
multidict==6.3.2
mypy-extensions==1.0.0
numpy==2.2.1
outcome==1.3.0.post0
packaging==24.2
parsimonious==0.10.0
pillow==11.2.1
playwright==1.51.0
pluggy==1.5.0
prompt-toolkit==3.0.36
propcache==0.3.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
pyee==12.1.1
Pygments==2.19.1
pyparsing==3.2.0
PySocks==1.7.1
pysui==0.81.0
pysui-fastcrypto==0.6.0
pytest==8.3.5
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pyunormalize==16.0.0
pyuseragents==1.0.5
pywin32==310
PyYAML==6.0.2
questionary==2.1.0
readchar==4.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rlp==4.1.0
rpds-py==0.24.0
runs==1.2.2
selenium==4.30.0
selenium-actions==1.0.6
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.40
tenacity==9.1.2
text-unidecode==1.3
toolz==1.0.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
types-requests==2.32.0.20250328
typing-inspect==0.9.0
typing-inspection==0.4.0
typing-utils==0.1.0
typing_extensions==4.13.0
urllib3==2.3.0
wcwidth==0.2.13
web3==7.10.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==11.0.3
win32_setctime==1.2.0
wrapt==1.17.2
wsproto==1.2.0
xmod==1.8.1
yarl==1.18.3
